+++ src/Fabulous.AST/Widgets/ModuleDeclarations/TopLevelBindings/TopLevelBinding.fs
@@ -96,3 +96,9 @@ type ValueYieldExtensions =
         let moduleDecl = ModuleDecl.TopLevelBinding node
         let widget = Ast.EscapeHatch(moduleDecl).Compile()
         { Widgets = MutStackArray1.One(widget) }
+
+    [<Extension>]
+    static member inline Yield(_: CollectionBuilder<'parent, ModuleDecl>, node: BindingNode) : CollectionContent =
+        let moduleDecl = ModuleDecl.TopLevelBinding node
+        let widget = Ast.EscapeHatch(moduleDecl).Compile()
+        { Widgets = MutStackArray1.One(widget) }